+++ src/system/schedule/sealed.rs
@@ -6,32 +6,22 @@ use crate::{
     },
 };
 
-pub trait Sealed<
-    'a,
-    R,
-    Resources,
-    I,
-    P,
-    RI,
-    MergeParametersList,
-    ResourcesIndicesLists,
-    ResourcesContainmentsLists,
-    ResourcesInverseIndicesLists,
-    QueryIndicesLists,
-    ResourceViewsIndicesLists,
-    DisjointIndicesLists,
-    EntryIndicesLists,
-> where
+pub trait Sealed<'a, R, Resources, Indices>
+where
     R: Registry,
 {
+    type QueryIndicesLists;
+    type ResourceViewsIndicesLists;
+    type DisjointIndicesLists;
+    type EntryIndicesLists;
     type Stages: Stages<
         'a,
         R,
         Resources,
-        QueryIndicesLists,
-        ResourceViewsIndicesLists,
-        DisjointIndicesLists,
-        EntryIndicesLists,
+        Self::QueryIndicesLists,
+        Self::ResourceViewsIndicesLists,
+        Self::DisjointIndicesLists,
+        Self::EntryIndicesLists,
     >;
 
     fn as_stages(&'a mut self) -> Self::Stages;
@@ -58,17 +48,19 @@ impl<
         'a,
         R,
         Resources,
-        I,
-        P,
-        RI,
-        MergeParametersList,
-        ResourcesIndicesLists,
-        ResourcesContainmentsLists,
-        ResourcesInverseIndicesLists,
-        QueryIndicesLists,
-        ResourceViewsIndicesLists,
-        DisjointIndicesLists,
-        EntryIndicesLists,
+        (
+            I,
+            P,
+            RI,
+            MergeParametersList,
+            ResourcesIndicesLists,
+            ResourcesContainmentsLists,
+            ResourcesInverseIndicesLists,
+            QueryIndicesLists,
+            ResourceViewsIndicesLists,
+            DisjointIndicesLists,
+            EntryIndicesLists,
+        ),
     > for T
 where
     R: Registry,
@@ -89,6 +81,10 @@ where
         EntryIndicesLists,
     >,
 {
+    type QueryIndicesLists = QueryIndicesLists;
+    type ResourceViewsIndicesLists = ResourceViewsIndicesLists;
+    type DisjointIndicesLists = DisjointIndicesLists;
+    type EntryIndicesLists = EntryIndicesLists;
     type Stages = <T as Scheduler<
         'a,
         R,