+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - `System` and `ParSystem` both no longer require `Filter` and `Views` to implement `Filter`.
 - `result::Iter` no longer requires `V` and `F` to implement `Filter`.
 - `Entry::query()` no longer requires `V` and `F` to implement `Filter`.
+- `system::schedule::Schedule` now only requires a single parameter for indices.
+- `World::run_system()` now only requires a single parameter for schedule indices.
 
 ## 0.8.2 - 2023-04-02
 ### Fixed